syntax/encoding: rename riverjson package to alloyjson
